gradio_demo/test/pyproject.toml
[project]
name = "larp-vlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "gradio>=5.35.0",
    "openai>=1.93.0",
    "pillow>=9.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "zhipuai>=2.0.0",
]
